program/API/sample/src/requirements.txt
aws-wsgi==0.2.7
Cerberus==1.3.2
cffi==1.15.0
click==8.0.3
Flask==2.0.2
httptools==0.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
pycryptodome==3.12.0
PyJWT==2.3.0
uvloop==0.16.0
websockets==10.1
Werkzeug==2.0.2
